Improve styling of progress bar
