+++ game/configs/mission/ui/inc/macroExecs.inc
@@ -138,5 +138,6 @@ __EXEC(\
 */
 #define VGM_SIZEEX_TT2020BASE(SIZEPX) __EVAL([SIZEPX,1.8,[6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,34,35,37,46]] call _fnc_VGMsizeEx)
 
+#define VGM_FONT_S VGM_SIZEEX_TT2020BASE(SIZEEX_S)
 #define VGM_FONT_M VGM_SIZEEX_TT2020BASE(SIZEEX_M)
 #define VGM_FONT_L VGM_SIZEEX_TT2020BASE(SIZEEX_L)
